Allow saving copy of english lines in an useful case:
When the option to save english text on localization files is enabled, an internal condition to succeed was that an english line already existed on destination file. Now, such copies are always saved if a translation is marked added/changed for a certain token on save. This is almost needed to ensure the english diffing functionality covers these affected tokens the future user work. Otherwise, user may miss to translate externally updated english lines.